Add a way to handle comics that use ISBN
Some comics use ISBNs instead of ISIN codes, Google Books has
those comics in its DB. Though the number of the comic is not
available and they all have the same name, to work around this
issue we check every time we save a book if there's another one
with the same name in our local DB already.
If one is found we open a dialog so the user can input the number
of the comic, we save that number in the DB afterward.